auto-merge: agent: resolve issue #223 (#509)
auto-merge: agent: resolve issue #223 (#509)
